add updateKeyword
